Improve implementation
