+++ src/views/components/database/trainer/editors/TrainerDialogEditor/useTrainerDialog.ts
@@ -10,9 +10,7 @@ import { useTrainerPage } from '@src/hooks/usePage';
 import { cloneEntity } from '@utils/cloneEntity';
 import { findFirstAvailableTextId } from '@utils/ModelUtils';
 import { useSetProjectText } from '@utils/ReadingProjectText';
-import { TFunction } from 'i18next';
 import { useMemo, useState } from 'react';
-import { useTranslation } from 'react-i18next';
 
 const getNewCondition = (dialogs: TrainerDialogAdditionalDialogs[]) => {
   const currentConditions = dialogs.map(({ condition }) => condition);
@@ -31,15 +29,6 @@ const getNewTextId = (additionalDialogs: StudioTrainerAdditionalDialogs[], dialo
   return findFirstAvailableTextId(allAdditionalDialogs);
 };
 
-const dialogConditionEntries = (dialogs: TrainerDialogAdditionalDialogs[], dialogIndex: number, t: TFunction<'database_trainers'>) => {
-  const currentConditions = dialogs.map(({ condition }) => condition).filter((condition) => condition !== dialogs[dialogIndex].condition);
-  const conditions = TRAINER_ADDITIONAL_DIALOGS_CONDITION.filter((condition) => !currentConditions.includes(condition)).map((condition) => ({
-    value: condition.toString(),
-    label: t(`additional_dialog_${condition}`),
-  }));
-  return conditions;
-};
-
 export type TrainerDialogAdditionalDialogs = {
   condition: 'default' | StudioTrainerAdditionalDialogs['condition'];
   textId: number;
@@ -47,7 +36,6 @@ export type TrainerDialogAdditionalDialogs = {
 
 export const useTrainerDialog = () => {
   const { trainer, trainers } = useTrainerPage();
-  const { t } = useTranslation('database_trainers');
   const setText = useSetProjectText();
   const allAdditionalDialogs = useMemo(() => getAllAdditionalDialogs(trainers, trainer), [trainers, trainer]);
   const [dialogIndex, setDialogIndex] = useState<number>(0);
@@ -58,7 +46,6 @@ export const useTrainerDialog = () => {
     },
     ...trainer.additionalDialogs,
   ]);
-  const [conditionOptions, setConditionOptions] = useState(dialogConditionEntries(dialogs, dialogIndex, t));
   const currentDialog = dialogs[dialogIndex];
 
   const addDialog = () => {
@@ -72,7 +59,6 @@ export const useTrainerDialog = () => {
     setTimeout(() => {
       setDialogs(newDialogs);
       setDialogIndex(newIndex);
-      setConditionOptions(dialogConditionEntries(newDialogs, newIndex, t));
     }, 0);
   };
 
@@ -86,7 +72,6 @@ export const useTrainerDialog = () => {
     dialogsUpdated.splice(index, 1);
     setDialogs(dialogsUpdated);
     updateDialogIndex(index - 1);
-    setConditionOptions(dialogConditionEntries(dialogsUpdated, index - 1, t));
   };
 
   const changeCondition = (condition: string) => {
@@ -97,7 +82,6 @@ export const useTrainerDialog = () => {
 
   const updateDialogIndex = (newIndex: number) => {
     setDialogIndex(newIndex);
-    setConditionOptions(dialogConditionEntries(dialogs, newIndex, t));
   };
 
   return {
@@ -105,7 +89,6 @@ export const useTrainerDialog = () => {
     currentDialog,
     dialogIndex,
     canAddDialog: dialogs.length < TRAINER_ADDITIONAL_DIALOGS_CONDITION.length + 1,
-    conditionOptions,
     updateDialogIndex,
     addDialog,
     deleteDialog,
